Refactor txHelper
